+++ src/Surfnet/StepupSelfService/SelfServiceBundle/Form/Type/PromiseSafeStorePossessionType.php
@@ -1,0 +1,56 @@
+<?php
+
+/**
+ * Copyright 2022 SURFnet bv
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace Surfnet\StepupSelfService\SelfServiceBundle\Form\Type;
+
+use Surfnet\StepupSelfService\SelfServiceBundle\Command\PromiseSafeStorePossessionCommand;
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class PromiseSafeStorePossessionType extends AbstractType
+{
+    public function buildForm(FormBuilderInterface $builder, array $options)
+    {
+        $builder->add('promisedPossession', CheckboxType::class, [
+            'label' => 'ss.form.recovery_token.checkbox.promise_possession',
+            'required' => true,
+            'attr' => [
+                'autofocus' => true,
+            ]
+        ])
+        ->add('confirm', SubmitType::class, [
+            'label' => 'ss.form.recovery_token.button.confirm',
+            'attr' => [ 'class' => 'btn btn-primary pull-right' ],
+        ]);
+    }
+
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver->setDefaults([
+            'data_class' => PromiseSafeStorePossessionCommand::class,
+        ]);
+    }
+
+    public function getBlockPrefix()
+    {
+        return 'ss_promise_recovery_token_possession';
+    }
+}