+++ src/Surfnet/StepupSelfService/SelfServiceBundle/Security/Authentication/SamlInteractionProvider.php
@@ -103,11 +103,11 @@ class SamlInteractionProvider
             $this->serviceProvider
         );
 
-        if ($assertion->getIssuer() !== $this->identityProvider->getEntityId()) {
+        if ($assertion->getIssuer()->getValue() !== $this->identityProvider->getEntityId()) {
             throw new UnexpectedIssuerException(sprintf(
                 'Expected issuer to be configured remote IdP "%s", got "%s"',
                 $this->identityProvider->getEntityId(),
-                $assertion->getIssuer()
+                $assertion->getIssuer()->getValue()
             ));
         }